Refactoring complete
Refactored AI execution tree of static class calls into instanced components.

TankAIHelper now always has a MovementController that handles all movement concerns.

The MovementController currently is either Air, or Default.

Each MovementController has a IMovementAICore which handles the actual process of setting and maintaining throttle/target locations.

The MovementController is responsible for selecting the correct AI Core - e.g. the AIControllerAir will swap between AirplaneAICore, VtolAICore, and HelicopterAICore on initialization.

AIControllerDefault will only use VehicleAICore (ground)
