+++ TAC_AI/AI/IMovementAIController.cs
@@ -1,0 +1,40 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using TAC_AI.AI.MovementAI;
+
+namespace TAC_AI.AI {
+
+    public interface IMovementAIController
+    {
+        IMovementAICore AICore
+        {
+            get;
+        }
+
+        Tank Tank
+        {
+            get;
+        }
+
+        AIECore.TankAIHelper Helper
+        {
+            get;
+        }
+
+        Enemy.RCore.EnemyMind EnemyMind
+        {
+            get;
+        }
+
+        void Initiate(Tank tank, AIECore.TankAIHelper helper, Enemy.RCore.EnemyMind mind = null);
+
+        void DriveDirector();
+
+        void DriveMaintainer(TankControl tankControl);
+
+        void Recycle();
+    }
+}
